requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.1
defusedxml==0.8.0rc2
diff-match-patch==20230430
Django==5.1.1
django-admin-display==1.3.0
django-import-export==4.1.1
django-jazzmin==3.0.0
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
drf-yasg==1.21.7
idna==3.10
inflection==0.5.1
oauthlib==3.2.2
packaging==24.1
pillow==10.4.0
pycparser==2.22
PyJWT==2.9.0
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rest-framework-simplejwt==0.0.2
setuptools==75.1.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
tablib==3.5.0
uritemplate==4.1.1
urllib3==2.2.3
